Update.
1999-07-18  Andreas Jaeger  <aj@arthur.rhein-neckar.de>

	* nscd/nscd_proto.h (__nscd_gethostbyaddr_r): Follow change from
	1999-07-17 and adjust len parameter.
	* nscd/nscd_gethst_r.c (__nscd_gethostbyaddr_r): Likewise.
	* nscd/gethstbyad_r.c (ADD_PARAMS): Likewise.
	* manual/socket.texi (Host Names): Likewise.
	* nis/nss_nis/nis-hosts.c (_nss_nis_gethostbyaddr_r): Likewise.
	* nis/nss_nisplus/nisplus-hosts.c (_nss_nisplus_gethostbyaddr_r):
	Likewise.
	* resolv/nss_dns/dns-host.c (_nss_dns_gethostbyaddr_r): Likewise.
	* resolv/gethnamaddr.c (gethostbyaddr): Likewise.
	(ht_gethostbyaddr): Likewise.
	(_gethtbyaddr): Likewise.

+++ nis/nss_nis/nis-hosts.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1996, 1997, 1998 Free Software Foundation, Inc.
+/* Copyright (C) 1996, 1997, 1998, 1999  Free Software Foundation, Inc.
    This file is part of the GNU C Library.
    Contributed by Thorsten Kukuk <kukuk@vt.uni-paderborn.de>, 1996.
 
@@ -341,7 +341,7 @@ _nss_nis_gethostbyname_r (const char *name, struct hostent *host, char *buffer,
 }
 
 enum nss_status
-_nss_nis_gethostbyaddr_r (char *addr, int addrlen, int type,
+_nss_nis_gethostbyaddr_r (char *addr, size_t addrlen, int type,
 			  struct hostent *host, char *buffer, size_t buflen,
 			  int *errnop, int *h_errnop)
 {
